Refactoring containerd code into its own package
