--- pkg/pillar/cmd/domainmgr/containerd.go
@@ -1,121 +1,0 @@
-// Copyright (c) 2020 Zededa, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-package domainmgr
-
-import (
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"path"
-	"path/filepath"
-	"strings"
-
-	v1 "github.com/opencontainers/image-spec/specs-go/v1"
-	uuid "github.com/satori/go.uuid"
-	log "github.com/sirupsen/logrus"
-)
-
-const (
-	// We look for <uuid>/image-config.json in here
-	// root path to all containers
-	containersRoot = "/persist/runx/pods/prepared"
-	// container config file name
-	imageConfigFilename = "image-config.json"
-)
-
-// getContainerPath return the path to the root of the container. This is *not*
-// necessarily the rootfs, which may be a layer below
-func getContainerPath(containerID string) string {
-	return path.Join(containersRoot, containerID)
-}
-
-func getImageInfo(containerID string) (v1.Image, error) {
-	var image v1.Image
-
-	appDir := getContainerPath(containerID)
-	data, err := ioutil.ReadFile(filepath.Join(appDir, imageConfigFilename))
-	if err != nil {
-		return image, err
-	}
-	if err := json.Unmarshal(data, &image); err != nil {
-		return image, err
-	}
-	return image, nil
-}
-
-// ctrRun eventually will run a container. For now, we do not actually run it via containerd,
-// but instead use containerd to set it up, and xl to run it, so this does not need to do anything.
-// If/when we invert it, and have containerd launch the container with a microvm wrapper,
-// *then* this will need to work.
-func ctrRun(domainName, xenCfgFilename, imageHash string, envList map[string]string) (int, string, error) {
-
-	log.Infof("ctrRun %s\n", domainName)
-	return 0, "", nil
-}
-
-// ctrStop eventually will stop a container. For now, we do not actually run it via containerd,
-// but instead use containerd to set it up, and xl to run it, so this does not need to do anything.
-// If/when we invert it, and have containerd launch the container with a microvm wrapper,
-// *then* this will need to work.
-func ctrStop(containerID string, force bool) error {
-	log.Infof("ctrStop %s %t\n", containerID, force)
-	log.Infof("ctr stop done\n")
-	return nil
-}
-
-func ctrPrepareMount(containerID uuid.UUID, containerPath string, envVars map[string]string, noOfDisks int) error {
-	log.Infof("ctrPrepareMount(%s, %s, %v, %d)", containerID, containerPath,
-		envVars, noOfDisks)
-	imageInfo, err := getImageInfo(containerID.String())
-	if err != nil {
-		log.Errorf("ctrPrepareMount(%s, %s) getImageInfo failed: %s",
-			containerID, containerPath, err)
-		return err
-	}
-	// inject a few files of our own into the bundle
-	mountpoints, execpath, workdir, env, err := getContainerConfigs(imageInfo, envVars)
-	if err != nil {
-		log.Errorf("ctrPrepareMount(%s, %s) getContainerConfigs failed: %s",
-			containerID, containerPath, err)
-		return fmt.Errorf("ctrPrepare: unable to get container config: %v", err)
-	}
-
-	err = createMountPointExecEnvFiles(containerPath, mountpoints, execpath, workdir, env, noOfDisks)
-	if err != nil {
-		log.Errorf("ctrPrepareMount(%s, %s) createMountPointExecEnvFiles failed: %s",
-			containerID, containerPath, err)
-	}
-	return err
-}
-
-// getContainerConfigs get the container configs needed, specifically
-// - mount target paths
-// - exec path
-// - working directory
-// - env var key/value pairs
-// this can change based on the config format
-func getContainerConfigs(imageInfo v1.Image, userEnvVars map[string]string) (map[string]struct{}, []string, string, []string, error) {
-
-	mountpoints := imageInfo.Config.Volumes
-	execpath := imageInfo.Config.Entrypoint
-	execpath = append(execpath, imageInfo.Config.Cmd...)
-	workdir := imageInfo.Config.WorkingDir
-	unProcessedEnv := imageInfo.Config.Env
-	var env []string
-	for _, e := range unProcessedEnv {
-		keyAndValueSlice := strings.Split(e, "=")
-		if len(keyAndValueSlice) == 2 {
-			//handles Key=Value case
-			env = append(env, fmt.Sprintf("%s=\"%s\"", keyAndValueSlice[0], keyAndValueSlice[1]))
-		} else {
-			//handles Key= case
-			env = append(env, e)
-		}
-	}
-
-	for k, v := range userEnvVars {
-		env = append(env, fmt.Sprintf("%s=\"%s\"", k, v))
-	}
-	return mountpoints, execpath, workdir, env, nil
-}
